refactor(dispatcher): разнести TelegramCommandDispatcher по доменам команд (#1047)
Горячая зона: src/services/telegram_command_dispatcher.py была монолитом на
1162 строки с MI rank C (0.00) — широкая mutation-поверхность (33 tool-команды,
phone-bound мутации, очередь команд). Тир-1 риск из карты сервисов #1022 (#12).

ПОЧЕМУ mixin, а не register-функции (как messaging_*.py): диспетчеризация
команд идёт через `getattr(self, "_handle_<type>")`, т.е. lookup обходит MRO
инстанса. Mixin-классы по доменам сохраняют каждый handler на self без
glue-кода, и `patch.object(type(d), ...)` в тестах продолжает работать через
MRO concrete-класса. Поведение НЕ менялось — только расположение методов.

Что сделано:
- 47 `_handle_*` разнесены в src/services/dispatcher/*_mixin.py по доменам
  (auth/accounts/scheduler/dialogs/channels/notifications/photo/search/moderation).
- Facade telegram_command_dispatcher.py (1162→436 стр, MI C→A) держит loop-
  машинерию (_run_loop, _update_command_safely, _dispatch, _unwrap_result_payload)
  и публичный класс/контракт. Все внешние импорты (web/bootstrap, cli/settings)
  и константы re-export'ятся как прежде.
- 5 патч-чувствительных handler'ов оставлены на facade-классе намеренно: тесты
  патчат TelegramActionService/Notifier/NotificationTargetService/__file__ через
  namespace ЭТОГО модуля, поэтому их call-site должен резолвить символы здесь
  (_handle_dialogs_join, _handle_dialogs_download_media, _handle_notifications_test,
  _notification_target_service).
- Константы и TelegramCommandRetryLaterError вынесены в _constants.py/_errors.py
  (разрыв цикла facade↔mixins), facade их re-export'ит.

Контракт сохранён:
- 150 характеризационных тестов (test_telegram_command_dispatcher.py) НЕ менялись
  и зелёные — red→green якорь, что handler'ы делают то же.
- Новый test_telegram_command_dispatcher_structure.py фиксирует структурный
  контракт разбивки (47 команд резолвятся, mixin-композиция, патч-точки на facade).
- test_telegram_action_contract.py: новые mixin-модули добавлены в ENTRYPOINT_FILES,
  чтобы инвариант делегирования Telegram-действий через TelegramActionService
  покрывал и их.

Метрики: facade 1162→436 стр; MI rank C(0.00)→A; except Exception в facade 9→2;
весь подпакет (13 файлов) — 100% MI rank A. Ноль НОВЫХ mypy-ошибок (5 имевшихся
в монолите переехали в mixin'ы как есть). ruff/lint-imports чисты, полный сьют
9344 passed.
